Update Docker publishing workflow
Removed the path restriction in Docker workflow to allow changes in any Dockerfile within the project. This provides greater flexibility and makes the project's Docker setup easier to maintain. Additional context was also incorporated into the event printout, and a condition-based "latest" tag was introduced to the image tagging process.