Disable scheduled Docker image publishing
The update made to the Docker publishing workflow on GitHub Actions involves disabling the cron job that was originally scheduled to automatically publish the Docker image. Henceforth, the image will instead be published only upon pushes to the "main" branch, ensuring images are updated only when substantial code changes are made.